# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "安信合同管理系统 API",
    "description": "基于 Go + Gin + MySQL 的合同管理系统 RESTful API 文档",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "host": "localhost:8000",
  "basePath": "/",
  "produces": ["application/json"],
  "consumes": ["application/json"],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Token. Example: \"Bearer {token}\""
    }
  },
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "summary": "用户注册",
        "description": "注册新用户账号",
        "tags": ["用户管理"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["username", "password"],
              "properties": {
                "username": {"type": "string", "example": "testuser"},
                "email": {"type": "string", "example": "test@example.com"},
                "password": {"type": "string", "example": "password123"},
                "full_name": {"type": "string", "example": "张三"},
                "role": {"type": "string", "enum": ["user", "manager", "admin"], "default": "user"}
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "注册成功",
            "schema": {
              "type": "object",
              "properties": {
                "id": {"type": "integer"},
                "username": {"type": "string"},
                "email": {"type": "string"},
                "message": {"type": "string"}
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "schema": {"type": "object", "properties": {"error": {"type": "string"}}}
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "用户登录",
        "description": "使用用户名密码登录获取JWT Token",
        "tags": ["用户管理"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["username", "password"],
              "properties": {
                "username": {"type": "string", "example": "admin"},
                "password": {"type": "string", "example": "admin123"}
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "登录成功",
            "schema": {
              "type": "object",
              "properties": {
                "access_token": {"type": "string"},
                "token_type": {"type": "string", "example": "bearer"},
                "user_info": {
                  "type": "object",
                  "properties": {
                    "id": {"type": "integer"},
                    "username": {"type": "string"},
                    "email": {"type": "string"},
                    "full_name": {"type": "string"},
                    "role": {"type": "string"}
                  }
                }
              }
            }
          },
          "401": {
            "description": "用户名或密码错误",
            "schema": {"type": "object", "properties": {"error": {"type": "string"}}}
          }
        }
      }
    },
    "/api/auth/users": {
      "get": {
        "summary": "获取用户列表",
        "description": "分页获取所有用户信息",
        "tags": ["用户管理"],
        "security": [{"Bearer": []}],
        "parameters": [
          {"in": "query", "name": "skip", "type": "integer", "description": "跳过记录数", "default": 0},
          {"in": "query", "name": "limit", "type": "integer", "description": "返回记录数", "default": 20}
        ],
        "responses": {
          "200": {
            "description": "成功",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {"type": "integer"},
                  "username": {"type": "string"},
                  "email": {"type": "string"},
                  "full_name": {"type": "string"},
                  "role": {"type": "string"},
                  "is_active": {"type": "boolean"}
                }
              }
            }
          }
        }
      }
    },
...
    "/api/customers": {
      "get": {
        "summary": "获取客户列表",
        "tags": ["客户管理"],
        "security": [{"Bearer": []}],
        "parameters": [
          {"in": "query", "name": "skip", "type": "integer"},
          {"in": "query", "name": "limit", "type": "integer"},
          {"in": "query", "name": "type", "type": "string", "enum": ["customer", "supplier"]}
        ],
        "responses": {
          "200": {"description": "成功"}
        }
      },
      "post": {
        "summary": "创建客户",
        "tags": ["客户管理"],
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "in": "body",
            "schema": {
              "type": "object",
              "required": ["name", "code"],
              "properties": {
                "name": {"type": "string", "example": "测试客户"},
                "code": {"type": "string", "example": "C001"},
                "type": {"type": "string", "enum": ["customer", "supplier"]},
                "contact_person": {"type": "string"},
                "contact_phone": {"type": "string"},
                "contact_email": {"type": "string"},
                "credit_rating": {"type": "string"}
              }
            }
          }
        ],
        "responses": {
          "201": {"description": "创建成功"}
        }
      }
    },
...
    "/api/contracts": {
      "get": {
        "summary": "获取合同列表",
        "tags": ["合同管理"],
        "security": [{"Bearer": []}],
        "parameters": [
          {"in": "query", "name": "skip", "type": "integer"},
          {"in": "query", "name": "limit", "type": "integer"},
          {"in": "query", "name": "customer_id", "type": "integer"},
          {"in": "query", "name": "contract_type_id", "type": "integer"},
          {"in": "query", "name": "status", "type": "string"}
        ],
        "responses": {
          "200": {"description": "成功"}
        }
      },
      "post": {
        "summary": "创建合同",
        "tags": ["合同管理"],
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "in": "body",
            "schema": {
              "type": "object",
              "required": ["title", "customer_id", "contract_type_id"],
              "properties": {
                "title": {"type": "string"},
                "customer_id": {"type": "integer"},
                "contract_type_id": {"type": "integer"},
                "amount": {"type": "number"},
                "currency": {"type": "string", "default": "CNY"},
                "sign_date": {"type": "string", "format": "date"},
                "start_date": {"type": "string", "format": "date"},
                "end_date": {"type": "string", "format": "date"},
                "payment_terms": {"type": "string"},
                "content": {"type": "string"},
                "notes": {"type": "string"}
              }
            }
          }
        ],
        "responses": {
          "201": {"description": "创建成功"}
        }
      }
    },
...
    "/api/statistics": {
      "get": {
        "summary": "获取统计数据",
        "tags": ["审批管理"],
        "security": [{"Bearer": []}],
        "responses": {
          "200": {
            "description": "成功",
            "schema": {
              "type": "object",
              "properties": {
                "total_contracts": {"type": "integer"},
                "active_contracts": {"type": "integer"},
                "pending_contracts": {"type": "integer"},
                "completed_contracts": {"type": "integer"},
                "total_amount": {"type": "number"},
                "this_month_contracts": {"type": "integer"},
                "this_month_amount": {"type": "number"},
                "expiring_soon": {"type": "integer"}
              }
            }
          }
        }
      }
    },
...
  },
...
}